refactor : #245 <Coupon> kafka 인터페이스 생성 및 구현체 적용 향후 redisson 적용을 위한 의존성 추가
- kafka interface 생성 후 구현체 적용
- 향후 redisson 사용을 위한 의존성 추가
