+++ com.taken_seat.coupon-service/build.gradle
@@ -52,6 +52,7 @@ dependencies {
 
 	implementation 'org.springframework.kafka:spring-kafka'
 	testImplementation 'org.springframework.kafka:spring-kafka-test'
+	implementation 'org.redisson:redisson-spring-boot-starter:3.41.0'
 
 	// db
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
