+++ com.taken_seat.coupon-service/src/main/java/com/taken_seat/coupon_service/application/kafka/CouponToUserConsumer.java
@@ -1,0 +1,12 @@
+package com.taken_seat.coupon_service.application.kafka;
+
+import com.taken_seat.common_service.message.KafkaUserInfoMessage;
+import org.springframework.kafka.support.KafkaHeaders;
+import org.springframework.messaging.handler.annotation.Header;
+import org.springframework.messaging.handler.annotation.Payload;
+
+public interface CouponToUserConsumer {
+
+    KafkaUserInfoMessage consume(@Payload KafkaUserInfoMessage message,
+                                        @Header(KafkaHeaders.RECEIVED_KEY) String key);
+}
