+++ com.taken_seat.coupon-service/src/main/java/com/taken_seat/coupon_service/infrastructure/kafka/CouponToUserConsumerImpl.java
@@ -1,7 +1,8 @@
-package com.taken_seat.coupon_service.presentation.controller;
+package com.taken_seat.coupon_service.infrastructure.kafka;
 
 import com.taken_seat.common_service.message.KafkaUserInfoMessage;
-import com.taken_seat.coupon_service.application.service.KafkaProducerService;
+import com.taken_seat.coupon_service.application.kafka.CouponToUserConsumer;
+import com.taken_seat.coupon_service.application.kafka.CouponToUserConsumerService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.KafkaHeaders;
@@ -12,12 +13,12 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Slf4j
-public class KafkaConsumer {
+public class CouponToUserConsumerImpl implements CouponToUserConsumer {
 
-    private final KafkaProducerService kafkaProducerService;
+    private final CouponToUserConsumerService couponToUserConsumerService;
 
-    public KafkaConsumer(KafkaProducerService kafkaProducerService) {
-        this.kafkaProducerService = kafkaProducerService;
+    public CouponToUserConsumerImpl(CouponToUserConsumerService couponToUserConsumerService) {
+        this.couponToUserConsumerService = couponToUserConsumerService;
     }
 
     private static final String RESPONSE_TOPIC = "Issuance-of-coupons";
@@ -25,6 +26,7 @@ public class KafkaConsumer {
     private static final String RECEIVE_MESSAGE_KEY = "Partitions-of-coupons";
 
     @KafkaListener(groupId = "couponFIFO", topics = RESPONSE_TOPIC)
+    @Override
     @SendTo(RESPONSE_TOPIC_REPLY)
     public KafkaUserInfoMessage consume(@Payload KafkaUserInfoMessage message,
                                         @Header(KafkaHeaders.RECEIVED_KEY) String key) {
@@ -32,7 +34,7 @@ public class KafkaConsumer {
         try {
             if (key.equals(RECEIVE_MESSAGE_KEY)) {
                 log.info("userId: {}", message.getUserId());
-                return kafkaProducerService.producerMessage(message);
+                return couponToUserConsumerService.producerMessage(message);
             }
         } catch (Exception e) {
             log.error(e.getMessage(), e);
